Remove index.php from website (sometimes visible in seo results)
I searched for the medialibrary docs and found that this url (https://spatie.be/index.php/docs/laravel-medialibrary/v9/introduction on https://www.google.com/search?q=laravel+media+library) uses `index.php` in the search results. 

Based on a blogpost by @mattiasgeniar (https://ma.ttias.be/remove-index-php-from-the-url-in-laravel/), it's possible to handle this in the routeserviceprovider. If this isn't the way you want (in the application itself), it could also be done in nginx or something else.